KRACOEUS-7078: update sessionKey, mirror form package in impl (move rule)
