+++ coeus-code/src/main/java/org/kuali/coeus/common/impl/sponsor/form/SponsorFormsMaintenanceDocumentRule.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kuali.coeus.common.impl.sponsor;
+package org.kuali.coeus.common.impl.sponsor.form;
 
 import org.apache.commons.lang3.StringUtils;
 import org.kuali.coeus.common.framework.sponsor.form.SponsorForms;
