+++ coeus-code/src/main/java/org/kuali/coeus/common/impl/sponsor/form/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * Kuali Coeus Sponsor Form related implementations
+ * 
+ * 
+ * 
+ * @author Kuali Coeus Development Team
+ *
+ */
+package org.kuali.coeus.common.impl.sponsor.form;
