+++ coeus-code/src/main/resources/org/kuali/kra/datadictionary/docs/SponsorFormsMaintenanceDocument.xml
@@ -18,7 +18,7 @@
     
     
     
-    <property name="businessRulesClass" value="org.kuali.coeus.common.impl.sponsor.SponsorFormsMaintenanceDocumentRule"/>
+    <property name="businessRulesClass" value="org.kuali.coeus.common.impl.sponsor.form.SponsorFormsMaintenanceDocumentRule"/>
     <property name="documentTypeName" value="SponsorFormsMaintenanceDocument"/>
     <property name="documentAuthorizerClass" value="org.kuali.rice.kns.document.authorization.MaintenanceDocumentAuthorizerBase"/>
